version propre avec réajout de dashboard, retour sur MainActivity.java mobile_navigation.xml / ajout des annotations et méthodes en DAO. Attention tous les ids sont des long désormais

+++ app/src/main/java/fr/uha/moreau/musicsgiver/model/MusicienGroupeAssociation.java
@@ -33,10 +33,10 @@ public class MusicienGroupeAssociation {
         return mid;
     }
 
-    public void setGid(int gid) {
+    public void setGid(long gid) {
         this.gid = gid;
     }
-    public void setMid(int mid) {
+    public void setMid(long mid) {
         this.mid = mid;
     }
 }
